+++ packages/melos/bin/melos.dart
@@ -1,5 +1,16 @@
+import 'dart:io';
+
 import 'package:melos/src/command_runner.dart';
+import 'package:melos/src/common/exception.dart';
 
 void main(List<String> arguments) {
-  MelosCommandRunner().run(arguments);
+  try {
+    MelosCommandRunner().run(arguments);
+  } on MelosException catch (err) {
+    stderr.writeln(err.toString());
+    exitCode = 1;
+  } catch (err) {
+    exitCode = 1;
+    rethrow;
+  }
 }