+++ packages/melos/test/mock_workspace_fs.dart
@@ -17,6 +17,7 @@
 import 'dart:io';
 
 import 'package:path/path.dart';
+import 'package:pub_semver/pub_semver.dart';
 
 import 'mock_fs.dart';
 
@@ -30,12 +31,20 @@ Directory createMockWorkspaceFs({
   Iterable<String> workspacePackagesGlobs = const ['packages/**'],
   Iterable<MockPackageFs> packages = const [],
   bool setCwdToWorkspace = true,
+  bool? intellij,
 }) {
-  assert(IOOverrides.current is MockFs,
-      'Mock workspaces can only be created inside a mock filesystem');
+  assert(
+    IOOverrides.current is MockFs,
+    'Mock workspaces can only be created inside a mock filesystem',
+  );
 
   // Create a `melos.yaml`
-  _createMelosConfig(workspaceRoot, workspaceName, workspacePackagesGlobs);
+  _createMelosConfig(
+    workspaceRoot,
+    workspaceName,
+    workspacePackagesGlobs,
+    intellij: intellij,
+  );
 
   // Sythesize a "package" (enough to satisfy our test requirements) for each
   // entry in `packages`
@@ -56,15 +65,30 @@ Directory createMockWorkspaceFs({
 void _createMelosConfig(
   String workspaceRoot,
   String workspaceName,
-  Iterable<String> workspacePackagesGlobs,
-) {
-  File(join(workspaceRoot, 'melos.yaml'))
-    ..createSync(recursive: true)
-    ..writeAsStringSync('''
+  Iterable<String> workspacePackagesGlobs, {
+  required bool? intellij,
+}) {
+  final melosYaml = File(join(workspaceRoot, 'melos.yaml'));
+
+  melosYaml.createSync(recursive: true);
+
+  melosYaml.writeAsStringSync(
+    '''
 name: $workspaceName
 packages:
 ${_yamlStringList(workspacePackagesGlobs)}
-''');
+''',
+  );
+
+  if (intellij != null) {
+    melosYaml.writeAsStringSync(
+      '''
+ide:
+  intellij: $intellij
+''',
+      mode: FileMode.append,
+    );
+  }
 }
 
 void _createPackage(MockPackageFs package, String workspaceRoot) {
@@ -73,10 +97,17 @@ void _createPackage(MockPackageFs package, String workspaceRoot) {
   if (package.publishToNone) {
     pubspec.writeln('publish_to: none');
   }
-  pubspec.writeln('''
+
+  if (package.version != null) {
+    pubspec.writeln('version: ${package.version}');
+  }
+
+  pubspec.writeln(
+    '''
 dependencies:
 ${_yamlMap(package.dependencyMap, indent: 2)}
-''');
+''',
+  );
 
   File(join(workspaceRoot, package.path, 'pubspec.yaml'))
     ..createSync(recursive: true)
@@ -98,6 +129,7 @@ class MockPackageFs {
     required this.name,
     String? path,
     List<String>? dependencies,
+    this.version,
     this.publishToNone = false,
     bool generateExample = false,
   })  : _path = path,
@@ -106,6 +138,8 @@ class MockPackageFs {
 
   /// Name of the package (must be a valid Dart package name)
   final String name;
+
+  final Version? version;
 
   /// Workspace-root relative path
   String get path => _path ?? 'packages/$name';